# Conflicts:
#	src/main/java/ru/hookaorder/backend/feature/place/entity/PlaceEntity.java

+++ src/main/java/ru/hookaorder/backend/feature/JWT/service/JwtUtils.java
@@ -17,7 +17,7 @@ public final class JwtUtils {
         final JwtAuthentication jwtInfoToken = new JwtAuthentication();
         jwtInfoToken.setRoles(getRoles(claims));
         jwtInfoToken.setPhone(claims.getSubject());
-        jwtInfoToken.setUserId(claims.get("userId", Integer.class));
+        jwtInfoToken.setUserId(claims.get("userId", Long.class));
         return jwtInfoToken;
     }
